@@ -1,14 +1,13 @@
 name: opml
 description: A Dart package for parsing and building OPML documents.
-version: 0.1.0
+version: 0.2.0
 homepage: https://github.com/daniel-keogh/opml
 
 environment:
-  sdk: '>=2.8.1 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  meta: ^1.2.2
-  xml: ^3.6.1
+  xml: ^5.1.2
 
 dev_dependencies:
   pedantic: ^1.9.0
