@@ -1,7 +1,32 @@
+import 'dart:io';
+
 import 'package:opml/opml.dart';
 
-// Creates an OpmlDocument object and prints it to the screen as XML.
 void main() {
+  print('read example...\n');
+  readExample();
+
+  print('\nwrite example...\n');
+  writeExample();
+}
+
+/// Reads a local OPML file and prints some properties.
+void readExample() {
+  final xml = File('example/example.xml').readAsStringSync();
+
+  final doc = OpmlDocument.parse(xml);
+
+  print(doc.head.title! + ':');
+
+  for (var category in doc.body) {
+    for (var feed in category.children!) {
+      print('- ${feed.title} : ${feed.xmlUrl}');
+    }
+  }
+}
+
+/// Creates an OpmlDocument object and prints it to the screen as XML.
+void writeExample() {
   final body = <OpmlOutline>[
     OpmlOutline(
       text: 'World',
